current state of endpoint description
