+++ .gitignore
@@ -156,4 +156,6 @@ dist
 apps/swirl-docs/specs
 apps/swirl-docs/problem.yml
 apps/swirl-docs/shared.yml
+apps/swirl-docs/users.yml
+apps/swirl-docs/usergroups.yml
 apps/swirl-docs/src/documents/api/*
