+++ apps/swirl-docs/src/lib/navigation/src/data/apiSpecs.data.ts
@@ -2,5 +2,5 @@
 import { NavItem } from "../navigation.model";
 
 export const apiSpecsNavItems: NavItem[] = [
-  {"title":"Tasks","url":"/api-docs/tasks","isRoot":true,"children":[{"title":"Get all tasks","url":"/api-docs/tasks#get-all-tasks","description":"get","isRoot":false},{"title":"Get number of tasks","url":"/api-docs/tasks#get-number-of-tasks","description":"get","isRoot":false},{"title":"Get all task assignments","url":"/api-docs/tasks#get-all-task-assignments","description":"get","isRoot":false},{"title":"Get all task comments","url":"/api-docs/tasks#get-all-task-comments","description":"get","isRoot":false},{"title":"Get all task reactions","url":"/api-docs/tasks#get-all-task-reactions","description":"get","isRoot":false},{"title":"Create a new task","url":"/api-docs/tasks#create-a-new-task","description":"post","isRoot":false},{"title":"Create new task assignment","url":"/api-docs/tasks#create-new-task-assignment","description":"post","isRoot":false},{"title":"Create multiple task assignments","url":"/api-docs/tasks#create-multiple-task-assignments","description":"post","isRoot":false},{"title":"Create a task comment","url":"/api-docs/tasks#create-a-task-comment","description":"post","isRoot":false},{"title":"Create new task comment reaction","url":"/api-docs/tasks#create-new-task-comment-reaction","description":"post","isRoot":false},{"title":"Delete a task","url":"/api-docs/tasks#delete-a-task","description":"delete","isRoot":false},{"title":"Delete existing task assignment","url":"/api-docs/tasks#delete-existing-task-assignment","description":"delete","isRoot":false},{"title":"Delete multiple task assignments","url":"/api-docs/tasks#delete-multiple-task-assignments","description":"delete","isRoot":false},{"title":"Delete existing task comment","url":"/api-docs/tasks#delete-existing-task-comment","description":"delete","isRoot":false},{"title":"Delete task comment reaction","url":"/api-docs/tasks#delete-task-comment-reaction","description":"delete","isRoot":false},{"title":"Modify/Update existing task","url":"/api-docs/tasks#modify/update-existing-task","description":"patch","isRoot":false}],"specName":"tasks.yml"},{"title":"User Management API","url":"/api-docs/user-management-api","isRoot":true,"children":[{"title":"Search Users","url":"/api-docs/user-management-api#search-users","description":"get","isRoot":false},{"title":"Get User Details","url":"/api-docs/user-management-api#get-user-details","description":"get","isRoot":false},{"title":"List linkable identity providers","url":"/api-docs/user-management-api#list-linkable-identity-providers","description":"get","isRoot":false},{"title":"Get linked identities for user","url":"/api-docs/user-management-api#get-linked-identities-for-user","description":"get","isRoot":false},{"title":"Check if a password is configured for a user","url":"/api-docs/user-management-api#check-if-a-password-is-configured-for-a-user","description":"get","isRoot":false},{"title":"Check if authenticated user has password","url":"/api-docs/user-management-api#check-if-authenticated-user-has-password","description":"get","isRoot":false},{"title":"Get email address of authenticated user","url":"/api-docs/user-management-api#get-email-address-of-authenticated-user","description":"get","isRoot":false},{"title":"Create User","url":"/api-docs/user-management-api#create-user","description":"post","isRoot":false},{"title":"Lock User","url":"/api-docs/user-management-api#lock-user","description":"post","isRoot":false},{"title":"Unlock User","url":"/api-docs/user-management-api#unlock-user","description":"post","isRoot":false},{"title":"Link identity to user","url":"/api-docs/user-management-api#link-identity-to-user","description":"post","isRoot":false},{"title":"Set password for user","url":"/api-docs/user-management-api#set-password-for-user","description":"post","isRoot":false},{"title":"Send email requesting account actions","url":"/api-docs/user-management-api#send-email-requesting-account-actions","description":"post","isRoot":false},{"title":"Send emails to multiple users requesting account actions","url":"/api-docs/user-management-api#send-emails-to-multiple-users-requesting-account-actions","description":"post","isRoot":false},{"title":"Change password of authenticated user","url":"/api-docs/user-management-api#change-password-of-authenticated-user","description":"post","isRoot":false},{"title":"Change email address of authenticated user","url":"/api-docs/user-management-api#change-email-address-of-authenticated-user","description":"post","isRoot":false},{"title":"Update User","url":"/api-docs/user-management-api#update-user","description":"patch","isRoot":false},{"title":"Delete User","url":"/api-docs/user-management-api#delete-user","description":"delete","isRoot":false},{"title":"Delete multiple users","url":"/api-docs/user-management-api#delete-multiple-users","description":"delete","isRoot":false},{"title":"Delete linked identity","url":"/api-docs/user-management-api#delete-linked-identity","description":"delete","isRoot":false},{"title":"Delete password of user","url":"/api-docs/user-management-api#delete-password-of-user","description":"delete","isRoot":false}],"specName":"users.yml"}
+  {"title":"Posts","url":"/api-docs/posts","isRoot":true,"children":[{"title":"Report a comment.","url":"/api-docs/posts#report-a-comment","description":"post","isRoot":false}],"specName":"posts.yml"},{"title":"Tasks","url":"/api-docs/tasks","isRoot":true,"children":[{"title":"Get all tasks","url":"/api-docs/tasks#get-all-tasks","description":"get","isRoot":false},{"title":"Get number of tasks","url":"/api-docs/tasks#get-number-of-tasks","description":"get","isRoot":false},{"title":"Get task assignments","url":"/api-docs/tasks#get-task-assignments","description":"get","isRoot":false},{"title":"Get a summary for the task assignments.","url":"/api-docs/tasks#get-a-summary-for-the-task-assignments","description":"get","isRoot":false},{"title":"Get all task comments","url":"/api-docs/tasks#get-all-task-comments","description":"get","isRoot":false},{"title":"Get all task reactions","url":"/api-docs/tasks#get-all-task-reactions","description":"get","isRoot":false},{"title":"Create a new task","url":"/api-docs/tasks#create-a-new-task","description":"post","isRoot":false},{"title":"Create new task assignment","url":"/api-docs/tasks#create-new-task-assignment","description":"post","isRoot":false},{"title":"Marking a task assignment as finished","url":"/api-docs/tasks#marking-a-task-assignment-as-finished","description":"post","isRoot":false},{"title":"Create multiple task assignments","url":"/api-docs/tasks#create-multiple-task-assignments","description":"post","isRoot":false},{"title":"Create a task comment","url":"/api-docs/tasks#create-a-task-comment","description":"post","isRoot":false},{"title":"Create new task comment reaction","url":"/api-docs/tasks#create-new-task-comment-reaction","description":"post","isRoot":false},{"title":"Delete a task","url":"/api-docs/tasks#delete-a-task","description":"delete","isRoot":false},{"title":"Delete existing task assignment","url":"/api-docs/tasks#delete-existing-task-assignment","description":"delete","isRoot":false},{"title":"Delete multiple task assignments","url":"/api-docs/tasks#delete-multiple-task-assignments","description":"delete","isRoot":false},{"title":"Delete existing task comment","url":"/api-docs/tasks#delete-existing-task-comment","description":"delete","isRoot":false},{"title":"Delete task comment reaction","url":"/api-docs/tasks#delete-task-comment-reaction","description":"delete","isRoot":false},{"title":"Modify/Update existing task","url":"/api-docs/tasks#modify/update-existing-task","description":"patch","isRoot":false},{"title":"Update task comment reaction","url":"/api-docs/tasks#update-task-comment-reaction","description":"patch","isRoot":false}],"specName":"tasks.yml"}
 ];
